Изменил названия тасков (для соответствия с нумерацией)
