+++ Battle_Of_Two_K/MNK.py
@@ -1,7 +1,6 @@
 import numpy as np
 from sympy.abc import x
 import matplotlib.pyplot as plt
-import matplotlib.ticker as ticker
 
 # /////////////////////////////////////////////
 # Аппроксимация  методом  наименьших  квадратов
@@ -11,24 +10,9 @@ import matplotlib.ticker as ticker
 X = [-6, -5, -4, -2]
 Y = [36, 27, 19, 7]
 
-
-# Вариант другой (Юра)
-# X = [-5, -3, -1, 0]
-# Y = [19, 7, 1, 0]
-
-# Вариант другой (Влад)
-# X = [-4, -3, -2, 0]
-# Y = [15, 8, 3, 0]
-
-
-# Вариант из методички
+# Вариант из методички (рабортает, можно проверить)
 # X = [-2, -1, 0, 1, 2]
 # Y = [3, 4, 2, 1, 1]
-
-
-# Вариант 11 (Сергей)
-# X = [-6, -5, -4, -2]
-# Y = [30, 22, 16, 6]
 
 # /////////////////////////////////////////////
 # Ниже програмный код!
@@ -141,43 +125,17 @@ print('/' * 41)
 print(f'Квадратичная невязка: {squared_residual()}')
 print('/' * 41)
 
-# \\\\\\\\\\\\\\\\\\\\\\
-# Красивый вывод графика
-# \\\\\\\\\\\\\\\\\\\\\\
+# \\\\\\\\\\\\\\\\\\\\\\\
+# Красивый вывод графика:
+# \\\\\\\\\\\\\\\\\\\\\\\
 
 x = np.linspace(-15, 15, 100)
 y = res[0] * x ** 2 + res[1] * x + res[2]
-
 figure, axes = plt.subplots()
-
 axes.plot(x, y, color='Navy')  # Узнать про linewidth
-
 axes.grid()
-
-
 axes.scatter(X, Y, color='red')
-
 axes.set_title('Аппроксимация методом наименьших квадратов')
-# axes.set(xlabel='Ось абсцис, x', ylabel='Ось ординат, y')
-
-# axes.xaxis.set_major_locator(ticker.MultipleLocator(2.5))
-# axes.xaxis.set_minor_locator(ticker.MultipleLocator(1.25))
-#
-# axes.yaxis.set_major_locator(ticker.MultipleLocator(5))
-# axes.yaxis.set_minor_locator(ticker.MultipleLocator(2.5))
-#
-# axes.grid(which='major')
-# axes.grid(which='minor')
-#
-# figure.set_figwidth(10)
-# figure.set_figheight(15)
-#
-# axes = plt.gca()
-# axes.spines['left'].set_position('center')
-# axes.spines['bottom'].set_position('center')
-# axes.spines['top'].set_visible(False)
-# axes.spines['right'].set_visible(False)
-
 plt.show()
 
 input()
